requirement.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
google-generativeai==0.3.0
opencv-python==4.8.1.78
numpy==1.24.3
speechrecognition==3.10.0
gtts==2.3.2
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
Pillow==10.0.0
bcrypt==4.0.1
cryptography==41.0.3
python-pptx==0.6.21
python-dotenv==1.0.0
face-recognition==1.3.0
qrcode==7.4.2
Flask-Migrate==4.0.5
